gather inference stuff in one place
