+++ packages/core/src/http/types.ts
@@ -1,8 +1,8 @@
 import type { TRPCError } from '../error/TRPCError';
+import type { inferRouterContext, inferRouterError } from '../inference';
 import type { ProcedureType } from '../procedure';
 import type { AnyRouter } from '../router';
 import type { TRPCResponse } from '../rpc';
-import type { inferRouterContext, inferRouterError } from '../types';
 import type { Dict } from '../utilityTypes';
 
 export type HTTPHeaders = Dict<string[] | string>;
