--- packages/core/src/internals/getParseFn.ts
@@ -1,43 +1,0 @@
-import type { Parser } from '../parser';
-
-export type ParseFn<TType> = (value: unknown) => Promise<TType> | TType;
-
-export function getParseFn<TType>(procedureParser: Parser): ParseFn<TType> {
-  const parser = procedureParser as any;
-
-  if (typeof parser === 'function') {
-    // ParserCustomValidatorEsque
-    return parser;
-  }
-
-  if (typeof parser.parseAsync === 'function') {
-    // ParserZodEsque
-    return parser.parseAsync.bind(parser);
-  }
-
-  if (typeof parser.parse === 'function') {
-    // ParserZodEsque
-    // ParserValibotEsque (<= v0.12.X)
-    return parser.parse.bind(parser);
-  }
-
-  if (typeof parser.validateSync === 'function') {
-    // ParserYupEsque
-    return parser.validateSync.bind(parser);
-  }
-
-  if (typeof parser.create === 'function') {
-    // ParserSuperstructEsque
-    return parser.create.bind(parser);
-  }
-
-  if (typeof parser.assert === 'function') {
-    // ParserScaleEsque
-    return (value) => {
-      parser.assert(value);
-      return value as TType;
-    };
-  }
-
-  throw new Error('Could not find a validator fn');
-}
